feat(msg): Add tag add, remove, and query methods
NeoMutt's practice Rust rewrite touched a block of code manipulating the
tags so this commit implements the necessary wrappers. With this commit,
it's possible to:

 - Add a tag
 - Remove a tag
 - Freeze/thaw message for bulk operations
 - Getting a list of tags.

Since these are just wrappers around the raw C APIs, there's not much
to talk about.

It's important to note none of the functions propagate the notmuch
status. It requires more thought to cleanly implement.